+++ _test/test/build_script_invalidation_test.dart
@@ -12,25 +12,25 @@ import 'package:test/test.dart';
 import 'common/utils.dart';
 
 void main() {
-  group('Build script changes', () {
-    setUp(() async {
-      ensureCleanGitClient();
-      await startServer(ensureCleanBuild: true);
-      addTearDown(() => stopServer(cleanUp: true));
-    });
+  setUp(() async {
+    ensureCleanGitClient();
+    await startServer(ensureCleanBuild: true, buildArgs: ['lib']);
+    addTearDown(() => stopServer(cleanUp: true));
+  });
 
+  group('Build script changes', () {
     test('while serving prompt the user to restart', () async {
       var filePath = p.join('pkgs', 'provides_builder', 'lib', 'builders.dart');
-      var terminateLine =
-          nextStdOutLine('Terminating. No further builds will be scheduled');
+      var expectedLines = [
+        'Terminating builds due to build script update',
+        'Creating build script snapshot',
+        'Building new asset graph',
+      ];
+      expect(stdOutLines, isNotNull);
+      for (var line in expectedLines) {
+        expect(stdOutLines, emitsThrough(contains(line)));
+      }
       await replaceAllInFile(filePath, RegExp(r'$'), '// do a build');
-      await terminateLine;
-      await stopServer();
-      await startServer(extraExpects: [
-        () => nextStdOutLine(
-            'Invalidating asset graph due to build script update'),
-        () => nextStdOutLine('Building new asset graph'),
-      ]);
     });
 
     test('while not serving invalidate the next build', () async {
@@ -43,21 +43,24 @@ void main() {
       var filePath = p.join('pkgs', 'provides_builder', 'lib', 'builders.dart');
       await stopServer();
       await replaceAllInFile(filePath, RegExp(r'$'), '// do a build');
-      await startServer(extraExpects: [
+      await startServer(buildArgs: [
+        'lib'
+      ], extraExpects: [
         () => nextStdOutLine(
             'Invalidating asset graph due to build script update'),
+        () => nextStdOutLine('Creating build script snapshot'),
         () => nextStdOutLine('Building new asset graph'),
+        () => nextStdOutLine('Succeeded after'),
       ]);
-
       expect(await File(extraFilePath).exists(), isFalse,
           reason: 'The cache dir should get deleted when the build '
               'script changes.');
-    });
+    }, onPlatform: {'windows': const Skip('flaky on windows')});
 
     test('Invalid asset graph version causes a new full build', () async {
       await stopServer();
-      var assetGraph = assetGraphPathFor(
-          p.join('.dart_tool', 'build', 'entrypoint', 'build.dart'));
+      var assetGraph = assetGraphPathFor(p.url
+          .join('.dart_tool', 'build', 'entrypoint', 'build.dart.snapshot'));
       // Prepend a 1 to the version number
       await replaceAllInFile(assetGraph, '"version":', '"version":1');
 
@@ -67,15 +70,34 @@ void main() {
       await createFile(extraFilePath, 'bar');
       expect(await File(extraFilePath).exists(), isTrue);
 
-      await startServer(extraExpects: [
+      await startServer(buildArgs: [
+        'lib'
+      ], extraExpects: [
         () => nextStdOutLine(
             'Throwing away cached asset graph due to version mismatch.'),
         () => nextStdOutLine('Building new asset graph'),
+        () => nextStdOutLine('Succeeded after'),
       ]);
-
       expect(await File(extraFilePath).exists(), isFalse,
           reason: 'The cache dir should get deleted when the asset graph '
               'can\'t be parsed');
-    }, onPlatform: {'windows': const Skip('flaky on windows')});
+    });
+  });
+
+  test('Build config changes rerun but dont invalidate the build', () async {
+    var expectedLines = [
+      'Terminating builds due to _test:build.yaml update',
+      'Builds finished. Safe to exit',
+      'Running build completed',
+      'with 0 outputs',
+    ];
+    expect(stdOutLines, isNotNull);
+    for (var line in expectedLines) {
+      expect(stdOutLines, emitsThrough(contains(line)));
+    }
+    await replaceAllInFile('build.yaml', '''
+targets:''', '''
+# Test Edit
+targets:''');
   });
 }