+++ _test/test/common/utils.dart
@@ -14,6 +14,7 @@ Directory _toolDir = Directory(p.join('.dart_tool', 'build'));
 
 Process _process;
 Stream<String> _stdOutLines;
+Stream<String> get stdOutLines => _stdOutLines;
 
 final String _pubBinary = Platform.isWindows ? 'pub.bat' : 'pub';
 
@@ -39,9 +40,15 @@ Future<Null> startServer(
         {bool ensureCleanBuild,
         List<Function> extraExpects,
         List<String> buildArgs}) =>
-    _startServer(_pubBinary,
-        ['run', 'build_runner', 'serve']..addAll(buildArgs ?? const []),
-        ensureCleanBuild: ensureCleanBuild, extraExpects: extraExpects);
+    _startServer(
+        'dart',
+        [
+          '--packages=.packages',
+          p.join('..', 'build_runner', 'bin', 'build_runner.dart'),
+          'serve'
+        ],
+        ensureCleanBuild: ensureCleanBuild,
+        extraExpects: extraExpects);
 
 Future<ProcessResult> _runBuild(String command, List<String> args,
     {bool ensureCleanBuild}) async {