+++ build_runner/bin/build_runner.dart
@@ -8,6 +8,7 @@ import 'dart:isolate';
 
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
+import 'package:build_runner_core/build_runner_core.dart';
 import 'package:io/ansi.dart';
 import 'package:io/io.dart';
 import 'package:logging/logging.dart';
@@ -17,6 +18,9 @@ import 'package:stack_trace/stack_trace.dart';
 import 'package:build_runner/src/build_script_generate/build_script_generate.dart';
 import 'package:build_runner/src/entrypoint/runner.dart';
 import 'package:build_runner/src/logging/std_io_logging.dart';
+
+const scriptSnapshotLocation = '$scriptLocation.snapshot';
+final logger = Logger('Bootstrap');
 
 Future<Null> main(List<String> args) async {
   // Use the actual command runner to parse the args and immediately print the
@@ -56,13 +60,26 @@ Future<Null> main(List<String> args) async {
   if (commandName != _generateCommand) {
     logListener = Logger.root.onRecord.listen(stdIOLogListener());
   }
+
+  while ((exitCode = await generateAndRun(commandName, args)) ==
+      ExitCode.tempFail.code) {}
+  await logListener?.cancel();
+}
+
+/// Generates the build script, snapshots it if needed, and runs it
+///
+/// Returns the desired exit code from running the build script.
+Future<int> generateAndRun(String commandName, List<String> args) async {
   var buildScript = await generateBuildScript();
   var scriptFile = File(scriptLocation)..createSync(recursive: true);
   scriptFile.writeAsStringSync(buildScript);
   if (commandName == _generateCommand) {
     print(p.absolute(scriptLocation));
-    return;
+    return 0;
   }
+
+  var scriptExitCode = await _createSnapshotIfMissing();
+  if (scriptExitCode != 0) return scriptExitCode;
 
   var exitPort = ReceivePort();
   var errorPort = ReceivePort();
@@ -75,35 +92,57 @@ Future<Null> main(List<String> args) async {
     final trace = e[1] as String;
     stderr.writeln(error);
     stderr.writeln(Trace.parse(trace).terse);
-    if (exitCode == 0) exitCode = 1;
+    if (scriptExitCode == 0) scriptExitCode = 1;
   });
   try {
-    await Isolate.spawnUri(
-        Uri.file(p.absolute(scriptLocation)), args, messagePort.sendPort,
+    await Isolate.spawnUri(Uri.file(p.absolute(scriptSnapshotLocation)), args,
+        messagePort.sendPort,
         onExit: exitPort.sendPort, onError: errorPort.sendPort);
   } on IsolateSpawnException catch (e) {
-    print(red.wrap('Failed to launch the build script. '
+    logger.severe(
+        'Failed to launch the build script. '
         'This is likely due to a misconfigured builder definition. '
-        'See the generated script at $scriptLocation to find errors.'));
-    print(e);
+        'See the generated script at $scriptLocation to find errors.',
+        e);
     messagePort.sendPort.send(ExitCode.config.code);
     exitPort.sendPort.send(null);
   }
   StreamSubscription exitCodeListener;
   exitCodeListener = messagePort.listen((isolateExitCode) {
-    if (isolateExitCode is! int) {
+    if (isolateExitCode is int) {
+      scriptExitCode = isolateExitCode;
+    } else {
       throw StateError(
           'Bad response from isolate, expected an exit code but got '
           '$isolateExitCode');
     }
-    exitCode = isolateExitCode as int;
     exitCodeListener.cancel();
     exitCodeListener = null;
   });
   await exitPort.first;
   await errorListener.cancel();
-  await logListener?.cancel();
   await exitCodeListener?.cancel();
+
+  return scriptExitCode;
+}
+
+/// Creates a script snapshot for the build script.
+///
+/// Returns zero for success or a number for failure which should be set to the
+/// exit code.
+Future<int> _createSnapshotIfMissing() async {
+  var snapshotFile = File(scriptSnapshotLocation);
+  if (!await snapshotFile.exists()) {
+    await logTimedAsync(logger, 'Creating build script snapshot...', () async {
+      await Process.run(Platform.executable,
+          ['--snapshot=$scriptSnapshotLocation', scriptLocation]);
+    });
+    if (!await snapshotFile.exists()) {
+      logger.severe('Failed to snapshot build script $scriptLocation');
+      return ExitCode.software.code;
+    }
+  }
+  return 0;
 }
 
 const _generateCommand = 'generate-build-script';