+++ build_runner/test/generate/watch_test.dart
@@ -393,7 +393,10 @@ a:file://different/fake/pkg/path
           test('to the root package', () async {
             await writer.writeAsString(
                 AssetId('a', 'build.yaml'), '# New build.yaml file');
-            expect(await results.hasNext, isFalse);
+            expect(await results.hasNext, isTrue);
+            var next = await results.next;
+            expect(next.status, BuildStatus.failure);
+            expect(next.failureType, FailureType.buildConfigChanged);
             expect(logs.length, 1);
             expect(logs.first.message,
                 contains('Terminating builds due to a:build.yaml update'));
@@ -403,7 +406,10 @@ a:file://different/fake/pkg/path
             await writer.writeAsString(
                 AssetId('b', 'build.yaml'), '# New build.yaml file');
 
-            expect(await results.hasNext, isFalse);
+            expect(await results.hasNext, isTrue);
+            var next = await results.next;
+            expect(next.status, BuildStatus.failure);
+            expect(next.failureType, FailureType.buildConfigChanged);
             expect(logs.length, 1);
             expect(logs.first.message,
                 contains('Terminating builds due to b:build.yaml update'));
@@ -412,7 +418,10 @@ a:file://different/fake/pkg/path
           test('<package>.build.yaml', () async {
             await writer.writeAsString(
                 AssetId('a', 'b.build.yaml'), '# New b.build.yaml file');
-            expect(await results.hasNext, isFalse);
+            expect(await results.hasNext, isTrue);
+            var next = await results.next;
+            expect(next.status, BuildStatus.failure);
+            expect(next.failureType, FailureType.buildConfigChanged);
             expect(logs.length, 1);
             expect(logs.first.message,
                 contains('Terminating builds due to a:b.build.yaml update'));
@@ -435,7 +444,10 @@ a:file://different/fake/pkg/path
             await writer.writeAsString(
                 AssetId('a', 'build.yaml'), '# Edited build.yaml file');
 
-            expect(await results.hasNext, isFalse);
+            expect(await results.hasNext, isTrue);
+            var next = await results.next;
+            expect(next.status, BuildStatus.failure);
+            expect(next.failureType, FailureType.buildConfigChanged);
             expect(logs.length, 1);
             expect(logs.first.message,
                 contains('Terminating builds due to a:build.yaml update'));
@@ -445,7 +457,10 @@ a:file://different/fake/pkg/path
             await writer.writeAsString(
                 AssetId('b', 'build.yaml'), '# Edited build.yaml file');
 
-            expect(await results.hasNext, isFalse);
+            expect(await results.hasNext, isTrue);
+            var next = await results.next;
+            expect(next.status, BuildStatus.failure);
+            expect(next.failureType, FailureType.buildConfigChanged);
             expect(logs.length, 1);
             expect(logs.first.message,
                 contains('Terminating builds due to b:build.yaml update'));
@@ -472,7 +487,10 @@ a:file://different/fake/pkg/path
             await writer.writeAsString(
                 AssetId('a', 'build.yaml'), '# Edited build.yaml file');
 
-            expect(await results.hasNext, isFalse);
+            expect(await results.hasNext, isTrue);
+            var next = await results.next;
+            expect(next.status, BuildStatus.failure);
+            expect(next.failureType, FailureType.buildConfigChanged);
             expect(logs.length, 1);
             expect(logs.first.message,
                 contains('Terminating builds due to a:build.yaml update'));
@@ -492,7 +510,10 @@ a:file://different/fake/pkg/path
             await writer.writeAsString(AssetId('a', 'build.cool.yaml'),
                 '# Edited build.cool.yaml file');
 
-            expect(await results.hasNext, isFalse);
+            expect(await results.hasNext, isTrue);
+            var next = await results.next;
+            expect(next.status, BuildStatus.failure);
+            expect(next.failureType, FailureType.buildConfigChanged);
             expect(logs.length, 1);
             expect(logs.first.message,
                 contains('Terminating builds due to a:build.cool.yaml update'));